+++ neurocaps/_utils/_cap_internals/__init__.py
@@ -1,4 +1,5 @@
 from ._capgetter import _CAPGetter
 from ._cap2statmap import _cap2statmap
+from ._heatmap_utils import _create_display, _save_contents
 from ._create_labels import _create_node_labels
 from ._run_kmeans import _run_kmeans