+++ neurocaps/analysis/__init__.py
@@ -2,5 +2,6 @@ from .cap import CAP
 from .change_dtype import change_dtype
 from .merge import merge_dicts
 from .standardize import standardize
+from .transition_matrix import transition_matrix
 
-__all__=["CAP", "change_dtype", "merge_dicts", "standardize"]
+__all__=["CAP", "change_dtype", "merge_dicts", "standardize", "transition_matrix"]